Initial commit.
shoutbox.py contains a class that reads and writes to the UBB.threads
shoutbox

xmppbridge.py is an xmppbridge based on xmpppy that can send messages
from a shoutbox object to a configured xmpp account.

conf.py is a configuration reader object based on ConfigParser

shoutbridge.py is a script that combines shoutbox.py and xmppbridge.py

headstockbridge.py is a bridge object based on headstock, that doesn't
work properly due to disconnection problem.