NEW: Assassin's Greed now has a functioning Buy action.
The assassin will receive gold when fighting, and this gold can be
used to buy items from a merchant. Neither the gold or the items
are listed in the stats at the moment. Also, the items are still
completely useless.

Potions will be required for healing, splinterbombs will be possible
to use as a one-time insta-kill, smokebombs will allow the assassin
to escape death once and daggers will increase damage in a fight.
All items are one-time use only.